requirements.txt
imageio==2.6.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
